Bug Fixes
This update eliminated some bugs that were present in the code. Note
that in order to run the component the files must  be in the format
where the directories are labeled as Subject 1, Subject 2 etc. And
within each directory, the event files should be labeled as Subject 1
Instance (1) or Subject 1 Instance (2) etc.